+++ lua/codecompanion/strategies/chat/tools/cmd_runner.lua
@@ -1,7 +1,13 @@
+--[[
+*Command Runner Tool*
+This tool is used to run shell commands on your system. It can handle multple
+commands in the same XML block. All commands must be approved by you.
+--]]
+
 local config = require("codecompanion.config")
 
 local log = require("codecompanion.utils.log")
-local util = require("codecompanion.utils.util")
+local util = require("codecompanion.utils")
 local xml2lua = require("codecompanion.utils.xml.xml2lua")
 
 ---@class CmdRunner.ChatOpts