+++ lua/codecompanion/utils/init.lua
@@ -48,17 +48,10 @@ end
 ---@param t table
 ---@return boolean
 M.is_array = function(t)
-  if type(t) ~= "table" then
-    return false
+  if type(t) == "table" and type(t[1]) == "table" then
+    return true
   end
-  local i = 0
-  for _ in pairs(t) do
-    i = i + 1
-    if t[i] == nil then
-      return false
-    end
-  end
-  return true
+  return false
 end
 
 ---@param table table
@@ -78,7 +71,7 @@ M._noop = function() end
 ---@param name string
 ---@return nil
 M.set_dot_repeat = function(name)
-  vim.go.operatorfunc = "v:lua.require'codecompanion.utils.util'._noop"
+  vim.go.operatorfunc = "v:lua.require'codecompanion.utils'._noop"
   vim.cmd.normal({ args = { "g@l" }, bang = true })
   vim.go.operatorfunc = string.format("v:lua.require'codecompanion'.%s", name)
 end